requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
sqlparse==0.5.2
stripe==11.6.0
typing_extensions==4.13.0
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.9.0
